Extract Storage interface
